Added staging buffers to improve model index and vertex buffer storage

+++ app/CoreAppRenderer.cpp
@@ -62,7 +62,7 @@ void CoreAppRenderer::start() {
 			{{-.5f, .5f, -0.5f}, {.1f, .8f, .1f}},
 			{{.5f, -.5f, -0.5f}, {.1f, .8f, .1f}},
 	};
-	std::vector<uint16_t> cubeIndices = { 0,  1,  2,  0,  3,  1,  4,  5,  6,  4,  7,  5,  8,  9,  10, 8,  11, 9,
+	std::vector<uint32_t> cubeIndices = { 0,  1,  2,  0,  3,  1,  4,  5,  6,  4,  7,  5,  8,  9,  10, 8,  11, 9,
 	                                     12, 13, 14, 12, 15, 13, 16, 17, 18, 16, 19, 17, 20, 21, 22, 20, 23, 21 };
 	auto cubeModel = std::make_shared<Model>(cubeVertices, cubeIndices);
